+++ gnome-shell/mpm@kdeldycke.github.io/prefs.js
@@ -86,7 +86,7 @@ export default class MpmPreferences extends ExtensionPreferences {
         logo.set_pixel_size(48);
         aboutRow.add_prefix(logo);
         const link = Gtk.LinkButton.new_with_label(
-            'https://kdeldycke.github.io/meta-package-manager/gnome-shell.html',
+            'https://mpm.run/gnome-shell.html',
             _('Documentation'));
         link.set_valign(Gtk.Align.CENTER);
         aboutRow.add_suffix(link);